+++ game_patch/multi/server.h
@@ -11,3 +11,4 @@ void server_do_frame();
 bool check_server_chat_command(const char* msg, rf::Player* sender);
 bool server_is_saving_enabled();
 void server_reliable_socket_ready(rf::Player* player);
+bool server_weapon_items_give_full_ammo();