+++ game_patch/object/object_private.h
@@ -8,5 +8,6 @@ void trigger_apply_patches();
 void obj_do_patch();
 void monitor_do_patch();
 void entity_do_patch();
+void item_do_patch();
 void particle_do_patch();
 void obj_light_apply_patch();