+++ game_patch/rf/player/player.h
@@ -167,4 +167,5 @@ namespace rf
     static auto& player_render = addr_as_ref<void(Player*)>(0x004A2B30);
     static auto& player_render_held_corpse = addr_as_ref<void(Player* player)>(0x004A2B90);
     static auto& player_do_frame = addr_as_ref<void(Player*)>(0x004A2700);
+    static auto& player_make_weapon_current_selection = addr_as_ref<void(Player *player, int weapon_type)>(0x004A4980);
 }