+++ docs/CHANGELOG.md
@@ -35,6 +35,7 @@ Version 1.6.2 (not released yet)
 - Fix persona messages being displayed too low in Big HUD mode in vehicles
 - Fix rfl file extension association with Dash Editor not working properly for filenames containing space characters
 - Add `mute_all_players` and `mute_player` commands for hiding chat messages from specific players
+- Add server config option to give max ammo to the player when a weapon item is picked up
 
 Version 1.6.1 (released 2021-02-20)
 -----------------------------------